+++ Assets/Library/In_Progress/Node.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: c01de42eb1ebe5347a4bbfca22c42197
+timeCreated: 1509198725
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
