+++ Assets/Library/In_Progress/ShrubUtils.cs
@@ -1,0 +1,8 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public static class ShrubUtils  {
+
+    public static System.Random random = new System.Random(System.DateTime.Now.Millisecond);
+}
